+++ src/layer/layerSymbols.ts
@@ -1,5 +1,9 @@
 /**
  * Symbol to declare a layers name on its visualizations, e.g. ol.layer.Layer, Cesium.Cesium3DTileset*
  */
-// eslint-disable-next-line import/prefer-default-export
 export const vcsLayerName: unique symbol = Symbol('vcsLayerName');
+
+/**
+ * Symbol added to Cesium3DTilesets to suppress picking.
+ */
+export const allowPicking: unique symbol = Symbol('allowPicking');